overlays: default: Fix keyboard layout by setting it to 'us'
AFAIK, there is no 'en' keyboard layout, but the most common one is
'us', so change it to that. This fixes the following error:

```sh
console-setup.sh[400]: /usr/bin/ckbcomp: Can not find file "symbols/en" in any known directory
```

And is the reason why the `console-setup.service` is currently failing.